Backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
python-dotenv==1.0.1
pydantic==2.9.2
groq==0.11.0
httpx==0.27.2
python-multipart==0.0.12
pymupdf==1.24.10
pdfplumber==0.11.4
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
email-validator==2.2.0
